# [Storage] Optimize memory cache key creation and key format for some stores