This speeds up key creation by ~372x and eliminates all memory allocs (used by key creation) for some stores. It also optimizes the cache key format to be smaller.

I found this "low-hanging fruit" optimization last week while working on a bugfix PR:
- #7324

Previously, memory cache key format for transaction results was inefficient and creates needless allocations by encoding to hex. Same issue exists for transaction result index cache and block cache. Light transaction results used by AN also use the same cache key format.

This PR optimizes cache key format for multiple stores:
- cache key format:
  - old: `hex(blockID) + hex(txID)`
  - new: `blockID + txID`
- index cache key format:
  - old `hex(blockID) + hex(txIndex)`
  - new `blockID + txIndex`
- block cache key format:
  - old `hex(blockID)`
  - new `blockID`

Updated stores include:
- `TransactionResults`
- `TransactionResultErrorMessages`
- `LightTransactionResults`

Since cache keys are created for every read or insert of transaction results, etc., it is good to reduce memory allocs when it doesn't sacrifice speed.

Benchmark stats for creating cache key:
```
                                               │    old.log    │               new.log               │
                                               │    sec/op     │   sec/op     vs base                │
TransactionResultCacheKey/_create_cache_key-12   882.000n ± 8%   2.368n ± 0%  -99.73% (p=0.000 n=10)

                                               │  old.log   │              new.log              │
                                               │    B/op    │   B/op    vs base                 │
TransactionResultCacheKey/_create_cache_key-12   576.0 ± 0%   0.0 ± 0%  -100.00% (p=0.000 n=10)

                                               │  old.log   │               new.log               │
                                               │ allocs/op  │ allocs/op   vs base                 │
TransactionResultCacheKey/_create_cache_key-12   9.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
```